+++ http/points/points_parser.go
@@ -4,6 +4,7 @@ import (
 	"compress/gzip"
 	"context"
 	"errors"
+	"fmt"
 	"io"
 	"io/ioutil"
 	"time"
@@ -13,8 +14,6 @@ import (
 	"github.com/influxdata/influxdb/v2/kit/tracing"
 	"github.com/influxdata/influxdb/v2/models"
 	"github.com/opentracing/opentracing-go"
-	"go.uber.org/zap"
-	"istio.io/pkg/log"
 )
 
 var (
@@ -82,7 +81,7 @@ func (pw *Parser) parsePoints(ctx context.Context, orgID, bucketID influxdb.ID, 
 	span.LogKV("values_total", len(points))
 	span.Finish()
 	if err != nil {
-		log.Error("Error parsing points", zap.Error(err))
+		tracing.LogError(span, fmt.Errorf("error parsing points: %v", err))
 
 		code := influxdb.EInvalid
 		// TODO - backport these