+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 1. [21479](https://github.com/influxdata/influxdb/pull/21477): Accept `--input` instead of a positional arg in `influx restore`.
 1. [21479](https://github.com/influxdata/influxdb/pull/21477): Print error instead of panicking when `influx restore` fails to find backup manifests.
 1. [21485](https://github.com/influxdata/influxdb/pull/21485): Set last-modified time of empty shard directory to the directory's mod time instead of Unix epoch.
+1. [21499](https://github.com/influxdata/influxdb/pull/21499): Remove erroneous dependency on istio.
 1. [21501](https://github.com/influxdata/influxdb/pull/21501): Don't deadlock in `influx org members list` when an org has > 10 members.
 
 ## v2.0.6 [2021-04-29]